requirements.txt
openai-whisper
sounddevice
numpy>=1.21.0
scipy
pandas>=1.5.0
scikit-learn>=1.0.0
seaborn>=0.11.0
matplotlib>=3.5.0 